+++ src/abis/abis.test.ts
@@ -1,4 +1,4 @@
-import * as abisJson from "./index";
+import * as abisJson from "./json";
 import * as abisTs from "./ts";
 
 describe("abis", () => {
@@ -7,8 +7,7 @@ describe("abis", () => {
   });
   it.each(Object.keys(abisTs))("%s ABI matches JSON version", (abiTs) => {
     const tsAbis = abisTs as unknown as any; // Avoid TS errors
-    const abiName = abiTs.substring(0, abiTs.length - 1); // TS ABIs exports have "TS" appended to avoid name collisions
-    const stringifiedTsAbi = JSON.stringify(tsAbis[abiName]);
+    const stringifiedTsAbi = JSON.stringify(tsAbis[abiTs]);
 
     const jsonAbis = abisJson as unknown as any; // Avoid TS errors
     const stringifiedJsonAbi = JSON.stringify(jsonAbis[abiTs]);
