+++ src/abis/json/CurrencyManagerAbi.json
