+++ src/abis/ts/index.ts
@@ -1,23 +1,23 @@
-export { default as AggregatorFeeSharingWithUniswapV3AbiTs } from "./AggregatorFeeSharingWithUniswapV3";
-export { default as CurrencyManagerAbiTs } from "./CurrencyManagerAbi";
-export { default as ERC1155AbiTs } from "./ERC1155";
-export { default as ERC20AbiTs } from "./ERC20";
-export { default as ERC721AbiTs } from "./ERC721";
-export { default as ExecutionManagerAbiTs } from "./ExecutionManager";
-export { default as FeeSharingSystemAbiTs } from "./FeeSharingSystem";
-export { default as IExecutionStrategyAbiTs } from "./IExecutionStrategy";
-export { default as LooksRareAirdropAbiTs } from "./LooksRareAirdrop";
-export { default as LooksRareExchangeAbiTs } from "./LooksRareExchange";
-export { default as Multicall2AbiTs } from "./Multicall2";
-export { default as MultiRewardsDistributorAbiTs } from "./MultiRewardsDistributor";
-export { default as OrderValidatorV1AbiTs } from "./OrderValidatorV1";
-export { default as PrivateSaleWithFeeSharingAbiTs } from "./PrivateSaleWithFeeSharing";
-export { default as ReverseRecordsAbiTs } from "./ReverseRecords";
-export { default as RoyaltyFeeManagerAbiTs } from "./RoyaltyFeeManager";
-export { default as RoyaltyFeeRegistryAbiTs } from "./RoyaltyFeeRegistry";
-export { default as RoyaltyFeeSetterAbiTs } from "./RoyaltyFeeSetter";
-export { default as StakingPoolForUniswapV2TokensAbiTs } from "./StakingPoolForUniswapV2Tokens";
-export { default as TokenDistributorAbiTs } from "./TokenDistributor";
-export { default as TradingRewardsDistributorAbiTs } from "./TradingRewardsDistributor";
-export { default as TransferSelectorNFTAbiTs } from "./TransferSelectorNFT";
-export { default as WETHAbiTs } from "./WETH";
+export { default as AggregatorFeeSharingWithUniswapV3Abi } from "./AggregatorFeeSharingWithUniswapV3";
+export { default as CurrencyManagerAbi } from "./CurrencyManagerAbi";
+export { default as ERC1155Abi } from "./ERC1155";
+export { default as ERC20Abi } from "./ERC20";
+export { default as ERC721Abi } from "./ERC721";
+export { default as ExecutionManagerAbi } from "./ExecutionManager";
+export { default as FeeSharingSystemAbi } from "./FeeSharingSystem";
+export { default as IExecutionStrategyAbi } from "./IExecutionStrategy";
+export { default as LooksRareAirdropAbi } from "./LooksRareAirdrop";
+export { default as LooksRareExchangeAbi } from "./LooksRareExchange";
+export { default as Multicall2Abi } from "./Multicall2";
+export { default as MultiRewardsDistributorAbi } from "./MultiRewardsDistributor";
+export { default as OrderValidatorV1Abi } from "./OrderValidatorV1";
+export { default as PrivateSaleWithFeeSharingAbi } from "./PrivateSaleWithFeeSharing";
+export { default as ReverseRecordsAbi } from "./ReverseRecords";
+export { default as RoyaltyFeeManagerAbi } from "./RoyaltyFeeManager";
+export { default as RoyaltyFeeRegistryAbi } from "./RoyaltyFeeRegistry";
+export { default as RoyaltyFeeSetterAbi } from "./RoyaltyFeeSetter";
+export { default as StakingPoolForUniswapV2TokensAbi } from "./StakingPoolForUniswapV2Tokens";
+export { default as TokenDistributorAbi } from "./TokenDistributor";
+export { default as TradingRewardsDistributorAbi } from "./TradingRewardsDistributor";
+export { default as TransferSelectorNFTAbi } from "./TransferSelectorNFT";
+export { default as WETHAbi } from "./WETH";
