+++ src/index.ts
@@ -1,5 +1,3 @@
-export * from "./abis";
-export * from "./abis/ts";
 export * from "./constants";
 export * from "./sign";
 export * from "./types";
